Use uint16_t for hi-speed interrupt variable type
This is needed to track wheel revolutions on nootnoot, as the encoder output of the motor board specifies 1000 interrupts == 3 revolutions, meaning that for tracking one full revolution, we need a maximum value of at least 333. uint8_t can only store values up to 255, so we need to change this here.